reduce storage

+++ contracts/popsicle-v3-optimizer/PopsicleV3Optimizer_arbitrum.sol
@@ -127,7 +127,7 @@ contract PopsicleV3Optimizer is ERC20Permit, ReentrancyGuard, IPopsicleV3Optimiz
     address public constant weth = 0x82aF49447D8a07e3bd95BD0d56f35241523fBab1;
     // @inheritdoc IPopsicleV3Optimizer
     int24 public immutable override tickSpacing;
-    uint256 constant MULTIPLIER = 1e6;
+    uint24 constant MULTIPLIER = 1e6;
     uint24 constant GLOBAL_DIVISIONER = 1e6; // for basis point (0.0001%)
     //The protocol's fee in hundredths of a bip, i.e. 1e-6
     uint24 constant protocolFee = 2 * 1e5; //20%
